requirements.txt
apipkg==1.4
astroid==1.5.3
bs4==0.0.1
dateparser==0.6.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.0.3
https://github.com/LexPredict/datefinder/archive/master.zip
dateparser==0.6.0
elasticsearch==5.4.0
execnet==1.4.1
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
lxml==4.2.0
mccabe==0.6.1
memory-profiler==0.47
nltk==3.2.4
nose==1.3.7
num2words==0.5.5
numpy==1.13.3
pandas==0.21.0
pep8==1.7.0
psutil==5.4.0
py==1.4.34
pycountry==17.1.8
pylint==1.7.2
pytest==3.2.1
pytest-cache==1.0
pytest-cov==2.5.1
pytest-pep8==1.0.6
pytest-pylint==0.7.1
pytest-xdist==1.20.1
python-coveralls==2.9.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
regex==2017.9.23
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
scikit-learn==0.19.1
tika==1.15
Unidecode==0.4.21
urllib3==1.22
us==1.0.0
wrapt==1.10.11
typing==3.6.2
flask-restful
pymongo
reporters_db
python-docx